+++ config/autoload/services.global.php
@@ -1,24 +1,22 @@
 <?php
 use Acelaya\Expressive\Factory\SlimRouterFactory;
 use Acelaya\Website\Action\Contact;
-use Acelaya\Website\Action\Factory\ActionAbstractFactory;
-use Acelaya\Website\Action\Factory\ContactFactory;
+use Acelaya\Website\Action\Template;
 use Acelaya\Website\Factory\CacheFactory;
 use Acelaya\Website\Factory\RecaptchaFactory;
 use Acelaya\Website\Factory\RendererFactory;
 use Acelaya\Website\Factory\RequestFactory;
 use Acelaya\Website\Factory\SwiftMailerFactory;
 use Acelaya\Website\Factory\TranslatorFactory;
+use Acelaya\Website\Form\ContactFilter;
 use Acelaya\Website\Middleware\CacheMiddleware;
-use Acelaya\Website\Middleware\Factory\CacheMiddlewareFactory;
-use Acelaya\Website\Middleware\Factory\LanguageMiddlewareFactory;
 use Acelaya\Website\Middleware\LanguageMiddleware;
 use Acelaya\Website\Options\Factory\MailOptionsFactory;
 use Acelaya\Website\Options\MailOptions;
 use Acelaya\Website\Service\ContactService;
-use Acelaya\Website\Service\Factory\ContactServiceFactory;
-use Acelaya\Website\Service\Factory\RouteAssemblerFactory;
+use Acelaya\Website\Service\ContactServiceInterface;
 use Acelaya\Website\Service\RouteAssembler;
+use Acelaya\ZsmAnnotatedServices\Factory\V3\AnnotatedFactory;
 use Doctrine\Common\Cache\Cache;
 use Psr\Http\Message\ServerRequestInterface;
 use ReCaptcha\ReCaptcha;
@@ -36,7 +34,8 @@ return [
             Application::class => ApplicationFactory::class,
 
             // Actions
-            Contact::class => ContactFactory::class,
+            Contact::class => AnnotatedFactory::class,
+            Template::class => AnnotatedFactory::class,
 
             // Services
             TemplateRendererInterface::class => RendererFactory::class,
@@ -46,24 +45,24 @@ return [
             RouterInterface::class => SlimRouterFactory::class,
             Cache::class => CacheFactory::class,
             ReCaptcha::class => RecaptchaFactory::class,
-            RouteAssembler::class => RouteAssemblerFactory::class,
-            ContactService::class => ContactServiceFactory::class,
+            RouteAssembler::class => AnnotatedFactory::class,
+            ContactService::class => AnnotatedFactory::class,
             'Zend\Expressive\FinalHandler' => TemplatedErrorHandlerFactory::class,
+            ContactFilter::class => AnnotatedFactory::class,
 
             // Options
             MailOptions::class => MailOptionsFactory::class,
 
             // Middleware
-            CacheMiddleware::class => CacheMiddlewareFactory::class,
-            LanguageMiddleware::class => LanguageMiddlewareFactory::class,
-        ],
-        'abstract_factories' => [
-            ActionAbstractFactory::class
+            CacheMiddleware::class => AnnotatedFactory::class,
+            LanguageMiddleware::class => AnnotatedFactory::class,
         ],
         'aliases' => [
             'translator' => Translator::class,
             'request' => ServerRequestInterface::class,
-            'renderer' => TemplateRendererInterface::class
+            'renderer' => TemplateRendererInterface::class,
+            ContactServiceInterface::class => ContactService::class,
+            AnnotatedFactory::CACHE_SERVICE => Cache::class,
         ]
     ]