+++ src/Service/RouteAssembler.php
@@ -1,6 +1,7 @@
 <?php
 namespace Acelaya\Website\Service;
 
+use Acelaya\ZsmAnnotatedServices\Annotation\Inject;
 use Psr\Http\Message\ServerRequestInterface;
 use Zend\Expressive\Router\RouteResult;
 use Zend\Expressive\Router\RouterInterface;
@@ -16,6 +17,13 @@ class RouteAssembler implements RouteAssemblerInterface
      */
     protected $request;
 
+    /**
+     * RouteAssembler constructor.
+     * @param RouterInterface $router
+     * @param ServerRequestInterface $request
+     *
+     * @Inject({RouterInterface::class, ServerRequestInterface::class})
+     */
     public function __construct(RouterInterface $router, ServerRequestInterface $request)
     {
         $this->router = $router;